# packages/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backline API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/search": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Search",
        "operationId": "search_api_v1_workspaces__workspace_id__search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "operationId": "list_projects_api_v1_workspaces__workspace_id__projects_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectOut"
                  },
                  "title": "Response List Projects Api V1 Workspaces  Workspace Id  Projects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "operationId": "create_project_api_v1_workspaces__workspace_id__projects_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_api_v1_projects__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update Project",
        "operationId": "update_project_api_v1_projects__project_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Archive Project",
        "operationId": "archive_project_api_v1_projects__project_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/restore": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Restore Project",
        "operationId": "restore_project_api_v1_projects__project_id__restore_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/settings": {
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update Project Settings",
        "description": "FD-AUD-018: persist the five review-settings flags for a project.",
        "operationId": "update_project_settings_api_v1_projects__project_id__settings_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectSettingsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSettingsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/duplicate": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Duplicate Project",
        "operationId": "duplicate_project_api_v1_projects__project_id__duplicate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/hard-delete/preview": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Preview Hard Delete Project",
        "operationId": "preview_hard_delete_project_api_v1_projects__project_id__hard_delete_preview_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectHardDeletePreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/hard-delete/confirm": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Confirm Hard Delete Project",
        "operationId": "confirm_hard_delete_project_api_v1_projects__project_id__hard_delete_confirm_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectHardDeleteConfirm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectHardDeleteResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/export": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Export Project",
        "operationId": "export_project_api_v1_projects__project_id__export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "target_origin": {
            "type": "string",
            "maxLength": 500,
            "title": "Target Origin",
            "default": ""
          },
          "project_type": {
            "type": "string",
            "enum": [
              "website",
              "image",
              "pdf"
            ],
            "title": "Project Type",
            "default": "website"
          },
          "environment": {
            "type": "string",
            "enum": [
              "live",
              "staging"
            ],
            "title": "Environment",
            "default": "live"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreate"
      },
...
      "ProjectHardDeleteConfirm": {
        "properties": {
          "correlation_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 36,
            "title": "Correlation Id"
          },
          "project_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Project Name"
          },
          "acknowledge_permanent_deletion": {
            "type": "boolean",
            "enum": [
              true
            ],
            "const": true,
            "title": "Acknowledge Permanent Deletion"
          }
        },
        "type": "object",
        "required": [
          "correlation_id",
          "project_name",
          "acknowledge_permanent_deletion"
        ],
        "title": "ProjectHardDeleteConfirm"
      },
      "ProjectHardDeletePreviewOut": {
        "properties": {
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "counts": {
            "$ref": "#/components/schemas/ProjectDeletionCounts"
          },
          "retention_notice": {
            "type": "string",
            "title": "Retention Notice"
          }
        },
        "type": "object",
        "required": [
          "correlation_id",
          "project_id",
          "project_name",
          "archived",
          "expires_at",
          "counts",
          "retention_notice"
        ],
        "title": "ProjectHardDeletePreviewOut"
      },
      "ProjectHardDeleteResult": {
        "properties": {
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "deleted"
            ],
            "const": "deleted",
            "title": "Status"
          },
          "counts": {
            "$ref": "#/components/schemas/ProjectDeletionCounts"
          }
        },
        "type": "object",
        "required": [
          "correlation_id",
          "status",
          "counts"
        ],
        "title": "ProjectHardDeleteResult"
      },
      "ProjectOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "target_origin": {
            "type": "string",
            "title": "Target Origin"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "settings": {
            "$ref": "#/components/schemas/ProjectSettingsOut"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "project_type": {
            "type": "string",
            "enum": [
              "website",
              "image",
              "pdf"
            ],
            "title": "Project Type",
            "default": "website"
          },
          "environment": {
            "type": "string",
            "enum": [
              "live",
              "staging"
            ],
            "title": "Environment",
            "default": "live"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "duplicated_from_project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicated From Project Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "name",
          "target_origin",
          "created_by",
          "settings",
          "archived_at",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectOut"
      },
      "ProjectSettingsOut": {
        "properties": {
          "proxy_mode": {
            "type": "boolean",
            "title": "Proxy Mode"
          },
          "snippet_installed": {
            "type": "boolean",
            "title": "Snippet Installed"
          },
          "capture_device_details": {
            "type": "boolean",
            "title": "Capture Device Details",
            "default": false
          },
          "reanchor_on_deploy": {
            "type": "boolean",
            "title": "Reanchor On Deploy",
            "default": false
          },
          "reviewer_can_resolve": {
            "type": "boolean",
            "title": "Reviewer Can Resolve",
            "default": false
          },
          "show_board_to_client": {
            "type": "boolean",
            "title": "Show Board To Client",
            "default": false
          },
          "client_digest_enabled": {
            "type": "boolean",
            "title": "Client Digest Enabled",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "proxy_mode",
          "snippet_installed"
        ],
        "title": "ProjectSettingsOut"
      },
      "ProjectSettingsUpdate": {
        "properties": {
          "capture_device_details": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capture Device Details"
          },
          "reanchor_on_deploy": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reanchor On Deploy"
          },
          "reviewer_can_resolve": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Can Resolve"
          },
          "show_board_to_client": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Board To Client"
          },
          "client_digest_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Digest Enabled"
          }
        },
        "type": "object",
        "title": "ProjectSettingsUpdate",
        "description": "Partial update for project review settings \u2014 omitted fields are unchanged."
      },
...
      "ProjectUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "target_origin": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Origin"
          },
          "environment": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "live",
                  "staging"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Environment"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          }
        },
        "type": "object",
        "title": "ProjectUpdate"
      },
...
      "SearchResultOut": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "project",
              "ticket",
              "comment",
              "member"
            ],
            "title": "Kind"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "subtitle": {
            "type": "string",
            "title": "Subtitle"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "page_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Id"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "id",
          "title",
          "subtitle"
        ],
        "title": "SearchResultOut",
        "description": "A deliberately small, route-safe projection for the shell search palette."
      },
      "SearchResultsOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchResultOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SearchResultsOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}